mv buildPayload to useBuildPayload hook
